+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Task display: Always show the sample init event in the task transcript display.
 - Inspect View: Outline view for transcript which enables high level navigation to solvers, agents, scorers, etc.
 - Inspect View: Fix an issue that prevented the display of the viewer in VSCode when the viewer tab was moved to the background.
+- Price limits: Prevent a sample from exceeding a specified USD-based budget.
 
 ## v0.3.99 (22 May 2025)